+++ saleor/graphql/product/types.py
@@ -139,6 +139,38 @@ class SelectedAttribute(graphene.ObjectType):
         description = 'Represents a custom attribute.'
 
 
+class DigitalContentUrl(CountableDjangoObjectType):
+    url = graphene.String(description='Url for digital content')
+
+    class Meta:
+        model = models.DigitalContentUrl
+        only_fields = ['token', 'content', 'created', 'download_num', 'url']
+        interfaces = (relay.Node,)
+
+    def resolve_url(self, info):
+        return self.get_absolute_url()
+
+
+class DigitalContent(CountableDjangoObjectType):
+    urls = gql_optimizer.field(
+        graphene.List(
+            lambda: DigitalContentUrl,
+            description='List of urls for the digital variant'),
+        model_field='urls')
+
+    class Meta:
+        model = models.DigitalContent
+        only_fields = [
+            'urls', 'content_file', 'use_default_settings',
+            'automatic_fulfillment', 'product_variant', 'max_downloads',
+            'url_valid_days', ]
+        interfaces = (relay.Node,)
+
+    def resolve_urls(self, info, **kwargs):
+        qs = self.urls.all()
+        return gql_optimizer.query(qs, info)
+
+
 class ProductOrder(graphene.InputObjectType):
     field = graphene.Argument(
         ProductOrderField, required=True,
@@ -185,6 +217,9 @@ class ProductVariant(CountableDjangoObjectType):
             'Returns translated Product Variant fields '
             'for the given language code.'),
         resolver=resolve_translation)
+    digital_content = gql_optimizer.field(graphene.Field(
+        DigitalContent, description='Digital content for the product variant'),
+        model_field='digital_content')
 
     class Meta:
         description = dedent("""Represents a version of a product such as
@@ -192,6 +227,10 @@ class ProductVariant(CountableDjangoObjectType):
         exclude_fields = ['order_lines', 'variant_images', 'translations']
         interfaces = [relay.Node]
         model = models.ProductVariant
+
+    @permission_required('product.manage_products')
+    def resolve_digital_content(self, info):
+        return getattr(self, 'digital_content', None)
 
     def resolve_stock_quantity(self, info):
         return self.quantity_available